Add bus crash signalisation + improve icons + gui improvements

+++ backend/src/main/java/vehicles/MovingObject.java
@@ -194,8 +194,7 @@ public abstract class MovingObject {
     // TODO: Sometimes index goes to 0
     public long getAdjacentOsmWayId(int indexFar) {
         int index = moveIndex + indexFar;
-        if(index >= uniformRoute.size())
-        {
+        if (index >= uniformRoute.size()) {
             return -1;
         }
         while (index >= 0 && !(uniformRoute.get(index) instanceof LightManagerNode)) {
@@ -227,7 +226,7 @@ public abstract class MovingObject {
 
     public int getMillisecondsToNextLight() {
         var distance = ((closestLightIndex - moveIndex) * RoutingConstants.STEP_CONSTANT);
-        return (int) (distance * CALCULATION_DELTA_PER_INDEX) + (int) (distance / getSpeed());
+        return (int) (distance * CALCULATION_DELTA_PER_INDEX) + (distance / getSpeed());
     }
 
     public int getMillisecondsFromAToB(int startIndex, int finishIndex) {
